implement timesheet time boundary checks
when deleting (clearing) timesheet or posting new entries, now there is
a check for the current timesheet's start and end dates, and if the
entry is outside of those bounds, nothing is done.

In the delete case, the script fully stops, because you need to talk to
your admin and deal with that shit. In the post case, it just skips it
with a warning.

I discovered an interesting thing today in that there's no enforcement
on the tenrox side for the start and end date of a timesheet. You can
enter time in future periods, but if you use this weeks timesheet id, it
will be linked to this week's timesheet (which the script loads at the
start based on the first time entry in your notes file)

So, you end up with entries that are not in the current timesheet, so it
doesn't impact this timesheet's time, but they get locked once this
timesheet is submitted and approved, and then you can't delete them when
they show up in that future timesheet.

This commit also has some formatting applied.