allow saturday time entries
corrected bug that was created when I started checking for entries outside of the current timesheet's date range. It was checking for entries beyond 00:00 on the final day of the timesheet. This now puts that date-time boundary at 23:59 (really 24:00) on that last day (saturday) of the time period, which will allow saturday time entries