support slashes in notes by using regex
to add year support in dates, we recently used a simple lookup for the date lines in the data, just looking for the slash character '/'. unfortunately I keep using slashes in my notes, so I've changed this back to using the regex... it's not fun for readability, but it lets me have slashes in my notes.